# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Captive Portal Omada/Radiusdesk API
  version: 0.1.0
  description: >-
    Backend API powering the Omada/Radiusdesk captive portal. All responses are JSON and
    instrumented for PM2 deployments.
servers:
  - url: http://localhost:4000
    description: Local dev
  - url: https://hotspot.techzone.lat
    description: Production
  - url: https://info.techzone.lat
    description: Production
...
security:
  - RadiusToken: []
components:
  securitySchemes:
    RadiusToken:
      type: apiKey
      in: query
      name: token
      description: RadiusDesk Access Provider token (paired with cloud_id query parameter).
    OmadaSession:
      type: apiKey
      in: cookie
      name: JSESSIONID
      description: Omada controller session cookie used after `/hotspot/login`.
    AdminToken:
      type: http
      scheme: bearer
      bearerFormat: uuid
      description: Operator session token issued by `/api/admin/login`.
    UsageSession:
      type: http
      scheme: bearer
      bearerFormat: usage-session
      description: Temporary bearer token issued by `/api/login` for end-user usage dashboards.
  schemas:
...
    RadacctSession:
      type: object
      properties:
        radacctid: { type: integer }
        username: { type: string }
        acctstarttime: { type: string, format: date-time }
        acctstoptime: { type: string, format: date-time, nullable: true }
        framedipaddress: { type: string }
...
    UsagePeriodSummary:
      type: object
      properties:
        period:
          type: string
          enum: [hourly, daily, weekly, monthly]
        totalBytes:
          type: number
          description: Aggregated input+output octets within the rolling window.
        totalTimeSeconds:
          type: number
          description: Total session duration in seconds within the window.
        sessionCount: { type: integer }
    UsageTimeseriesBucket:
      type: object
      properties:
        index:
          type: integer
          description: Zero-based position in the returned array.
        label:
          type: string
          description: Short label for the bucket (hour, weekday or day-of-month).
        start:
          type: string
          format: date-time
        end:
          type: string
          format: date-time
        totalBytes: { type: number }
        totalTimeSeconds: { type: number }
        sessionCount: { type: integer }
    UsageTimeseries:
      type: object
      properties:
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        granularity:
          type: string
          enum: [hour, day, month]
        buckets:
          type: array
          items:
            $ref: '#/components/schemas/UsageTimeseriesBucket'
...
    AdminAuthModes:
      type: object
      properties:
        availableModes:
          type: array
          items:
            type: string
            enum: [static, radiusmysql]
        defaultMode:
          type: string
          nullable: true
          enum: [static, radiusmysql, null]
    AdminLoginResult:
      type: object
      properties:
        token:
          type: string
          format: uuid
        expiresAt:
          type: integer
          format: int64
          description: Epoch milliseconds when the token expires.
        mode:
          type: string
          enum: [static, radiusmysql]
    AdminSession:
      type: object
      properties:
        username: { type: string }
        mode:
          type: string
          enum: [static, radiusmysql]
    RouterUsageStat:
      type: object
      properties:
        label: { type: string }
        totalBytes: { type: number }
        totalTimeSeconds: { type: number }
        sessionCount: { type: integer }
    AdminUserInsights:
      type: object
      properties:
        username: { type: string }
        macs:
          type: array
          items: { type: string }
        historyLimit: { type: integer }
        periods:
          type: array
          items:
            $ref: '#/components/schemas/UsagePeriodSummary'
        series:
          $ref: '#/components/schemas/UsageTimeseries'
        activeSessions:
          type: array
          items:
            $ref: '#/components/schemas/RadacctSession'
        inactiveSessions:
          type: array
          items:
            $ref: '#/components/schemas/RadacctSession'
        activeCount: { type: integer }
        inactiveCount: { type: integer }
        routerStats:
          type: array
          items:
            $ref: '#/components/schemas/RouterUsageStat'
        lastUpdated:
          type: string
          format: date-time
...
    ApiSuccess:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
    ApiError:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message: { type: string }
        status: { type: integer }
paths:
...
  /api/admin/auth-modes:
    get:
      tags: [admin]
      summary: List operator authentication modes
      description: Returns the configured admin authentication sources (static credentials or RadiusDesk MySQL user).
      operationId: listAdminAuthModes
      responses:
        '200':
          description: Available modes
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiSuccess'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/AdminAuthModes'
        '503':
          description: Admin authentication not configured
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /api/admin/login:
    post:
      tags: [admin]
      summary: Authenticate an operator and issue a bearer token
      operationId: loginAdmin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username: { type: string }
                password: { type: string, format: password }
                mode:
                  type: string
                  enum: [static, radiusmysql]
            examples:
              static:
                value:
                  username: admin
                  password: secret
                  mode: static
      responses:
        '200':
          description: Authenticated
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiSuccess'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/AdminLoginResult'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '503':
          description: Mode not available
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /api/admin/me:
    get:
      tags: [admin]
      summary: Retrieve the active admin session
      security:
        - AdminToken: []
      responses:
        '200':
          description: Current session
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiSuccess'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/AdminSession'
        '401':
          description: Missing or expired token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /api/admin/logout:
    post:
      tags: [admin]
      summary: Revoke the admin token
      security:
        - AdminToken: []
      responses:
        '200':
          description: Token removed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiSuccess'
        '401':
          description: Missing or expired token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /api/admin/users/{username}/insights:
    get:
      tags: [admin, usage]
      summary: Operator-only usage insights for a username
      description: Aggregates usage summaries, router stats and recent sessions without requiring the user password.
      operationId: getAdminUserInsights
      security:
        - AdminToken: []
      parameters:
        - in: path
          name: username
          required: true
          schema: { type: string }
        - in: query
          name: historyLimit
          schema:
            type: integer
            minimum: 50
            maximum: 500
          description: Optional override for the amount of historical sessions inspected.
      responses:
        '200':
          description: Usage insights
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiSuccess'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/AdminUserInsights'
        '401':
          description: Missing token or RadiusDesk authentication failure
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: Username not found or unable to determine MAC address
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '500':
          description: RadiusDesk timeout or aggregation failure
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
...
  /api/social/{provider}/start:
    get:
      tags: [social]
      summary: Begin the OAuth/OIDC handshake for dynamic social providers
      description: Returns the provider-specific authorization URL and state token.
      operationId: startSocialAuth
      parameters:
        - in: path
          name: provider
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Authorization URL
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiSuccess'
              examples:
                facebook:
                  value:
                    success: true
                    data:
                      provider: facebook
                      authorizationUrl: https://auth/facebook?state=req_123
        '400':
          description: Unsupported provider or missing parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: Social login disabled for the dynamic profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '500':
          description: Social provider bootstrap failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /api/social/{provider}/callback:
    get:
      tags: [social]
      summary: Callback endpoint storing state before connect/social
      description: Validates the OAuth callback and relays state back to the frontend flow.
      operationId: handleSocialCallback
      parameters:
        - in: path
          name: provider
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Callback acknowledged
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiSuccess'
        '400':
          description: Missing OAuth state or provider error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: Social login disabled for the dynamic profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '500':
          description: Internal callback processing error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
...